+++ zuul/layoutvalidator.py
@@ -55,6 +55,7 @@ class LayoutSchema(object):
                 'success-message': str,
                 'failure-message': str,
                 'dequeue-on-new-patchset': bool,
+                'dequeue-on-conflict': bool,
                 'trigger': toList(trigger),
                 'success': variable_dict,
                 'failure': variable_dict,